requirements.txt
numpy>=1.21.0
PyQt5>=5.15.0
scikit-learn>=0.24.0
matplotlib>=3.4.0
pillow>=8.0.0 